+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog: recurtools
 
+- **rename** `chainanything` to `starchain`, **default behaviour changed** to `recursive = True`
+- standardised on keywork argument `preserve`
+
 ## v0.4.0 - update to flatten (BREAKING)
 - **new keyword-only argument** `dontflatten` (optional): type which will not be flattened. Default: `(str, bytes)`.
 - `preservestrings` argument **removed**.